Extra useUtxo hook
