+++ packages/example/src/api/v1/activities.ts
