+++ packages/example/src/api/v1/coin.ts
