+++ packages/example/src/api/v1/register.ts
@@ -15,4 +15,3 @@ export const register = (
 ): Promise<RegisterResponse> => {
   return query(endpoint, RequestType.Post, {}, {channel: CHANNEL}, hdToken);
 }
-
