+++ packages/example/src/api/v1/sendInfo.ts
