+++ packages/example/src/api/v2/coin.ts
@@ -1,11 +1,11 @@
-import {RequestType} from "../types";
+import { RequestType } from "../types";
 import { query } from "../request-utils/query";
 import { SupportedCoins } from "../../constant/supportedCoins";
 
 const endpoint = '/v2/self_custody/coins/';
 
 export type CoinsResponse = {
-  coins: Record<SupportedCoins, {decimal: number; balance: string}>;
+  coins: Record<SupportedCoins, { decimal: number; balance: string }>;
 };
 
 export const queryCoinV2 = (): Promise<CoinsResponse> => {
