+++ packages/example/src/components/Account/Details.tsx
@@ -1,6 +1,11 @@
+import React from "react";
 import { Modal } from 'semantic-ui-react';
 import CloseIcon from "../Icons/CloseIcon";
 import { ReactComponent as AccountIcon } from "./image/wallet.svg";
+import { satoshiToBTC } from "../../lib/helper";
+import { bitcoinUnit } from "./Main";
+import { BitcoinNetwork } from "../../interface";
+import { useUtxo } from "../../hook/useUtxo";
 import {
   AccountDetailTop,
   ModalHeader,
@@ -11,29 +16,15 @@ import {
   AccountTopUnits,
   AccountDetailBottom, AccountListItem, AccountListLabel, AccountListLabelTop, AccountListLabelBottom
 } from "./styles"
-import React, { useEffect, useState } from "react";
-import { sendInfo } from "../../api/v1/sendInfo";
-import { useKeystoneStore } from "../../mobx";
-import { fromHdPathToObj } from "../../lib/cryptoPath";
-import { coinManager } from "../../services/CoinManager";
-import { satoshiToBTC } from "../../lib/helper";
-import { bitcoinUnit } from "./Main";
-import { BitcoinNetwork, Utxo } from "../../interface";
-
-interface CountedUtxo extends Utxo {
-  count: number;
-}
 
 type AccountDetails = {
-  open: boolean;
   close: () => void;
   balance: number;
   units: string;
 }
 
-const Details = (({open, close, balance, units}: AccountDetails) => {
-  const {current} = useKeystoneStore();
-  const [utxoList, setUtxoList] = useState<CountedUtxo[]>([]);
+const Details = (({close, balance, units}: AccountDetails) => {
+  const { utxoList } = useUtxo()
 
   const switchValue = ((value: number) => {
     if(units === bitcoinUnit[BitcoinNetwork.Main].BTC || units === bitcoinUnit[BitcoinNetwork.Test].BTC) {
@@ -43,44 +34,11 @@ const Details = (({open, close, balance, units}: AccountDetails) => {
     }
   })
 
-  useEffect(() => {
-    if (current && open) {
-      sendInfo(current.coinCode).then(data => {
-        const utxoList = data.spendables
-          .map(utxo => {
-            const {change, index} = fromHdPathToObj(utxo.hdPath);
-            const pubkey = coinManager.xpubToPubkey(current.xpub, Number(change || 0), Number(index || 0));
-            return {
-              transactionHash: utxo.txid,
-              index: Number(index || 0),
-              address: coinManager.deriveAddress(pubkey, current.scriptType, current.network),
-              value: utxo.value,
-              path: `M/${change}/${index}`,
-            }
-          }).reduce((acc: CountedUtxo[], cur): CountedUtxo[] => {
-            const itemIndex = acc.findIndex(item => item.path === cur.path);
-            if (itemIndex > -1) {
-              acc[itemIndex].count++;
-              acc[itemIndex].value += cur.value;
-              return acc;
-            } else {
-              return [...acc, {...cur, count: 1}]
-            }
-          }, [])
-        setUtxoList(utxoList)
-      }).catch(e => {
-        console.error("Fetch utxo list failed", e);
-      })
-    }
-  }, [open, current])
-
   return (
     <Modal
       className={'modal-container'}
-      open={open}
-      style={{
-        height: '640px'
-      }}
+      open={true}
+      style={{ height: 640 }}
     >
       <AccountDetailTop>
         <ModalHeader>
