+++ packages/example/src/components/TransactionCard/types.ts
