+++ packages/example/src/constant/bitcoin.ts
