+++ packages/example/src/constant/supportedCoins.ts
